sel4-microkit: Add support for embedded stack
