+++ crates/private/tests/microkit/reset/src/bin/test.rs
@@ -21,7 +21,7 @@ static mut PERSISTENT: usize = INIT;
 #[unsafe(link_section = ".persistent")]
 static mut RESET_COUNT: usize = 0;
 
-#[protection_domain]
+#[protection_domain(stack_size = 16 * 1024)]
 fn init() -> NullHandler {
     unsafe {
         debug_println!("NOT_PERSISTENT: {NOT_PERSISTENT}");
