+++ crates/sel4-runtime-common/src/start.rs
@@ -1,5 +1,5 @@
 //
-// Copyright 2023, Colias Group, LLC
+// Copyright 2025, Colias Group, LLC
 //
 // SPDX-License-Identifier: BSD-2-Clause
 //
@@ -7,31 +7,36 @@
 use core::arch::global_asm;
 
 #[macro_export]
-macro_rules! declare_stack {
-    ($size:expr) => {
-        const _: () = {
-            #[allow(non_upper_case_globals)]
-            #[unsafe(no_mangle)]
-            static __sel4_runtime_common__stack_bottom: $crate::_private::StackBottom = {
-                static STACK: $crate::_private::Stack<{ $size }> = $crate::_private::Stack::new();
-                STACK.bottom()
-            };
-        };
+macro_rules! declare_entrypoint {
+    () => {
+        $crate::_private::global_asm! {
+            r#"
+                .extern __sel4_runtime_common__call_rust_entrypoint
+
+                .section .text
+
+                .global _start
+                _start:
+            "#,
+            #[cfg(any(target_arch = "aarch64", target_arch = "arm"))]
+            r#"
+                    b __sel4_runtime_common__call_rust_entrypoint
+            "#,
+            #[cfg(any(target_arch = "riscv64", target_arch = "riscv32"))]
+            r#"
+                    j __sel4_runtime_common__call_rust_entrypoint
+            "#,
+            #[cfg(target_arch = "x86_64")]
+            r#"
+                    jmp __sel4_runtime_common__call_rust_entrypoint
+            "#,
+        }
     };
 }
 
 #[macro_export]
 macro_rules! declare_entrypoint_with_stack_init {
-    ($f:ident($( $i:ident: $t:ty ),* $(,)?)) => {
-        const _: () = {
-            #[unsafe(no_mangle)]
-            unsafe extern "C" fn __sel4_runtime_common__rust_entrypoint($($i: $t,)*) -> ! {
-                $crate::_private::_run_entrypoint(true, || {
-                    $f($($i,)*)
-                });
-            }
-        };
-
+    () => {
         $crate::_private::global_asm! {
             r#"
                 .extern __sel4_runtime_common__stack_init
@@ -55,6 +60,51 @@ macro_rules! declare_entrypoint_with_stack_init {
             "#,
         }
     };
+}
+
+#[macro_export]
+macro_rules! declare_stack {
+    ($size:expr) => {
+        const _: () = {
+            #[allow(non_upper_case_globals)]
+            #[unsafe(no_mangle)]
+            static __sel4_runtime_common__stack_bottom: $crate::_private::StackBottom = {
+                static STACK: $crate::_private::Stack<{ $size }> = $crate::_private::Stack::new();
+                STACK.bottom()
+            };
+        };
+    };
+}
+
+global_asm! {
+    r#"
+        .extern __sel4_runtime_common__rust_entrypoint
+
+        .section .text.__sel4_runtime_common__call_rust_entrypoint, "ax", %progbits
+
+        .global __sel4_runtime_common__call_rust_entrypoint
+        __sel4_runtime_common__call_rust_entrypoint:
+    "#,
+    #[cfg(any(target_arch = "aarch64", target_arch = "arm"))]
+    r#"
+            bl __sel4_runtime_common__rust_entrypoint
+        1:  b 1b
+    "#,
+    #[cfg(any(target_arch = "riscv64", target_arch = "riscv32"))]
+    r#"
+        .option push
+        .option norelax
+        1:  auipc gp, %pcrel_hi(__global_pointer$)
+            addi gp, gp, %pcrel_lo(1b)
+        .option pop
+            jal __sel4_runtime_common__rust_entrypoint
+        1:  j 1b
+    "#,
+    #[cfg(target_arch = "x86_64")]
+    r#"
+            call __sel4_runtime_common__rust_entrypoint
+        1:  jmp 1b
+    "#,
 }
 
 global_asm! {
